requirements.txt
agate==1.7.0
attrs==24.2.0
babel==2.16.0
certifi==2023.11.17
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dbt-core==1.5.11
dbt-extractor==0.4.1
dbt-sqlite==1.5.0
future==1.0.0
hologram==0.0.16
idna==3.8
isodate==0.6.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
leather==0.4.0
Logbook==1.5.3
MarkupSafe==2.1.5
mashumaro==3.6
minimal-snowplow-tracker==0.0.2
msgpack==1.1.0
nba_api==1.5.0
networkx==2.8.8
numpy==1.26.4
packaging==24.1
pandas==2.2.2
parsedatetime==2.4
pathspec==0.11.2
protobuf==4.25.4
pycparser==2.22
python-dateutil==2.8.2
python-dotenv==1.0.1
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
setuptools==75.0.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==2.3.8
